+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- BPay handling to NSW GovPay provider
+
 ## [10.1.2] - 2026-05-07
 
 ### Added